2 version

+++ src/main/java/com/example/bookstore/repository/BookRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BookRepo extends JpaRepository<Book, Integer> {
+public interface BookRepo extends JpaRepository<Book, Long> {
     List<Book> findAllByActiveTrue();
